refactor: separate into components separate components that load the room image modify type modify path from which a file is loaded
