+++ frontend/src/components/Room/Modal/Modal.style.ts
@@ -1,5 +1,29 @@
 import styled from "styled-components";
 
+const ModalBackground = styled.div`
+  position: fixed;
+  top:0;
+  left:0;
+  width:100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.6);
+  z-index: 99;
+`;
+
+const RoomAmenity = styled.div`
+  position: fixed;
+  top: 100px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 700px;
+  height: 700px;
+  background-color: #fff;
+  box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.4);
+  text-align: center;
+  overflow-y: scroll;
+  z-index: 100;
+`;
+
 const ModalHeaderWrapper = styled.div`
   position: sticky;
   top: 0px;
@@ -92,4 +116,4 @@ const SelectBtn = styled.button`
   cursor: pointer;
 `;
 
-export {ModalHeaderWrapper, CloseBtn, ModalBodyWrapper, RoomImage, ModalFooterWrapper, SelectBtn}
+export { ModalBackground, RoomAmenity, ModalHeaderWrapper, CloseBtn, ModalBodyWrapper, RoomImage, ModalFooterWrapper, SelectBtn }
