+++ frontend/src/components/Room/Room.tsx
@@ -1,10 +1,11 @@
 import { memo, useState } from 'react';
-import DetailRoomInfo from './DetailRoomInfo';
-import { Selector, RoomWrapper, Image } from './Room.style';
+import DetailRoomInfo from './Modal/DetailRoomInfo';
+import { Selector, RoomWrapper } from './Room.style';
 import { DetailRoomProps, RoomProps } from './Room.types';
+import { RoomImage } from './RoomImage';
 import { RoomInfo } from './RoomInfo';
 
-const Room = ({ room, setSelectedRoom }:{ room:DetailRoomProps, setSelectedRoom: (value: RoomProps) => void}) => {
+const Room = ({ room, setSelectedRoom }:{ room:DetailRoomProps, setSelectedRoom: (value: RoomProps) => void}):JSX.Element => {
   const [modal, setModal] = useState(false);
 
   const handleClick = ():void => {
@@ -21,13 +22,7 @@ const Room = ({ room, setSelectedRoom }:{ room:DetailRoomProps, setSelectedRoom:
       <DetailRoomInfo room={room} modal={modal} toggleModal={toggleModal} handleClick={handleClick} />
       <RoomWrapper onClick={toggleModal}>
         <Selector />
-        <Image>
-          {room.images[0] ? (
-            <img src={room.images[0].fullSizeUrl} alt={room.images[0].caption}></img>
-          ) : (
-            <img src={'https://img.icons8.com/ios/344/no-image.png'} style={{ width: '30px', height: '30px' }} alt="이미지 없음" ></img>
-          )}
-        </Image>
+        <RoomImage images={room.images} />
         <RoomInfo name={room.name} maxOccupancy={room.maxOccupancy} price={room.ratePlans[0].price} />
       </RoomWrapper>
     </li>
